Data Manipulation With R
This case is about Stack Overflow datasets. Stack Overflow is a forum where programmers around the world can discuss. The forum is in the form of a question and answer discussion, where anyone can give their opinion. The website accommodates all kinds of questions from all programming languages. The four datasets we will use are:
1. questions.csv → dataset of questions asked, storing column `id` as identifier.
2. answers.csv → dataset about the answers given, column `id` as the main identifier and has column `question_id` to combine against dataset `questions`.
3. tags.csv → dataset that defines a tag/question for a particular topic.
4. question_tags.csv → dataset that assigns each question to a specific tag.

In this case, we counted how many times each `question_id` appears in the `answers.csv` dataset. Joined the results to the dataset column `questions.csv`. Then, joined to dataset `questions_tags.csv` and to dataset `tags.csv`. From resulted, aggregation based on column `tag_name` to get the number of times the question appears. The last sorted the final result of the `tag_name` that occurs most often.